Add local caching system; adding Spring AOP to handle logging.

+++ src/main/java/com/fletch22/orb/cache/local/OrbSingleTypesInstanceCollection.java
@@ -10,9 +10,9 @@ import com.googlecode.cqengine.index.unique.UniqueIndex;
 public class OrbSingleTypesInstanceCollection {
 
 	IndexedCollection<CacheEntry> instances = new ConcurrentIndexedCollection<CacheEntry>();
-	private int orbTypeInternalId;
+	private long orbTypeInternalId;
 
-	public OrbSingleTypesInstanceCollection(int orbTypeInternalId) {
+	public OrbSingleTypesInstanceCollection(long orbTypeInternalId) {
 		this.orbTypeInternalId = orbTypeInternalId;
 		instances.addIndex(UniqueIndex.onAttribute(CacheEntry.ID));
 	}
